Kafka for JUnit has been - sort of - archived (#107)
Kafka for JUnit will not be updated regularly any longer. Unfortunately, I do not have the time to allocate my attention to this project in a sustainable fashion. Updates and bugfixes may still happen (support for new versions, ...), but if they do, it will be solely based on my personal requirements. Thank you for your understanding.